update neovim to use more threads. Update golang to only install when version doesn't match

+++ install/10-install-golang.zsh
@@ -3,6 +3,8 @@
 set -u
 set -o pipefail
 
+VERSION="1.17.2"
+
 echo "=============== INSTALLING GOLANG ==============="
 sleep 5
 
@@ -14,6 +16,12 @@ else
     echo "OKAY"
 fi
 
+echo "=============== checking current installation ==============="
+if [ -e $(command -v go) ] &&
+    [ $(go version | awk '/go/ {print $3}') = "go$VERSION" ]; then
+    echo "Current installation up to date!"
+    exit 0
+fi
 
 echo "=============== removing old install ==============="
 rm -rf $GOROOT
@@ -23,7 +31,7 @@ echo "OKAY"
 echo "=============== downloading archive ==============="
 mkdir tmp
 cd tmp
-curl -L https://golang.org/dl/go1.17.2.linux-amd64.tar.gz -o go1.17.2.linux-amd64.tar.gz
+curl -L "https://golang.org/dl/go$VERSION.linux-amd64.tar.gz" -o go1.17.2.linux-amd64.tar.gz
 echo "OKAY"
